初始化

+++ server.py
@@ -10,30 +10,31 @@ logger = logging.create_logger(app)
 
 
 @app.route('/')
-def hello_world():
-    return 'Hello World!'
+def root():
+    return 'godtoy\'s python api , Use WeChat：zhaojunlike to contact me'
 
 
 @app.route('/tx/image', methods=['POST'])
 def image():
-    json = request.get_json()
-    url = json['url']
-    # 1.下载图片
-    # 2.识别图片
-    app.logger.debug(url)
-    file = os.path.join(os.path.dirname(__file__), "tmp", str(uuid.uuid4()) + "_captcha.jpg")
-    print(file)
-    download_image_as_jpeg(url, "./tmp/a.jpg")
-    res = qq_mark_detect("./tmp/a.jpg")
-    dis = res.x.values[0]
-    tacks = get_track(dis)
-    print("解析成功：", dis)
     try:
-        os.remove(file)
-    except EnvironmentError as e:
+        json = request.get_json()
+        url = json['url']
+        # 1.下载图片
+        file = os.path.join(os.path.dirname(__file__), "tmp", str(uuid.uuid4()) + "_captcha.jpg")
+        _, code = download_image_as_jpeg(url, file)
+        if not code == 200:
+            return make_response(jsonify({'message': "解析失败Code:" + code}), 400)
+        # 2.识别图片
+        res = qq_mark_detect(file)
+        dis = res.x.values[0]
+        tacks = get_track(dis)  # 模拟加速度
+        app.logger.debug("解析成功,需要移动距离:{}".format(dis))
+        os.remove(file)  # 解析后删除文件就可以了
+        return make_response(jsonify({'message': "解析成功", 'data': {'x': dis, 'list': tacks, 'url': url}}), 200)
+    except Exception as e:
         print("e：", e)
-    return make_response(jsonify({'message': "解析成功", 'data': {'x': dis, 'list': tacks, 'url': url}}), 200)
+        return make_response(jsonify({'message': "解析失败,请重新尝试"}), 400)
 
 
 if __name__ == '__main__':
-    app.run(port=5000, host='0.0.0.0', debug=True)
+    app.run()
